requirements.txt
pyyaml
pandas
seaborn
wordcloud
pyyaml
matplotlib
torch
transformers